Format code

+++ db/migrate/20191107165911_add_missing_unique_indices.acts_as_taggable_on_engine.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # This migration comes from acts_as_taggable_on_engine (originally 2)
 class AddMissingUniqueIndices < ActiveRecord::Migration
   def self.up
@@ -5,18 +7,16 @@ class AddMissingUniqueIndices < ActiveRecord::Migration
       add_index :tags, :name, unique: true
     end
 
-    if index_exists? :taggings, :tag_id
-      remove_index :taggings, :tag_id
+    remove_index :taggings, :tag_id if index_exists? :taggings, :tag_id
+
+    if index_exists? :taggings, %i[taggable_id taggable_type context]
+      remove_index :taggings, %i[taggable_id taggable_type context]
     end
 
-    if index_exists? :taggings, [:taggable_id, :taggable_type, :context]
-      remove_index :taggings, [:taggable_id, :taggable_type, :context]
-    end
-
-    unless index_exists? :taggings, [:tag_id, :taggable_id, :taggable_type, :context, :tagger_id, :tagger_type], unique: true, name: 'taggings_idx'
-    add_index :taggings,
-              [:tag_id, :taggable_id, :taggable_type, :context, :tagger_id, :tagger_type],
-              unique: true, name: 'taggings_idx'
+    unless index_exists? :taggings, %i[tag_id taggable_id taggable_type context tagger_id tagger_type], unique: true, name: 'taggings_idx'
+      add_index :taggings,
+                %i[tag_id taggable_id taggable_type context tagger_id tagger_type],
+                unique: true, name: 'taggings_idx'
     end
   end
 
@@ -25,6 +25,6 @@ class AddMissingUniqueIndices < ActiveRecord::Migration
 
     remove_index :taggings, name: 'taggings_idx'
     add_index :taggings, :tag_id
-    add_index :taggings, [:taggable_id, :taggable_type, :context]
+    add_index :taggings, %i[taggable_id taggable_type context]
   end
 end
